+++ gaios/Utility/AssetSelect/AssetSelectCell/AssetSelectCell.swift
@@ -21,12 +21,25 @@ class AssetSelectCell: UITableViewCell {
 
     func configure(model: AssetSelectCellModel,
                    showEditIcon: Bool) {
+        ampSubview.isHidden = true
+        assetSubview.borderWidth = 0.0
+        iconEdit.isHidden = !showEditIcon
+        
+        // Any liquid asset
+        if model.anyLiquid {
+            self.lblAsset.text = "id_receive_any_liquid_asset".localized
+            self.imgView?.image = UIImage(named: "default_asset_liquid_icon")!
+            return
+        }
+        // Any AMP asset
+        if model.anyAmp {
+            self.lblAsset.text = "Receive any Amp Asset".localized
+            self.imgView?.image = UIImage(named: "default_asset_amp_icon")!
+            return
+        }
         let name = model.asset?.name ?? model.asset?.assetId
         self.lblAsset.text = name
         self.imgView?.image = model.icon
 
-        ampSubview.isHidden = true
-        assetSubview.borderWidth = 0.0
-        iconEdit.isHidden = !showEditIcon
     }
 }